requirements.txt
django===5.0.6
django-cors-headers
django-extensions===3.1.3
cryptography
django-cryptography
django-encrypted-model-fields
pycryptodome
requests
urllib3
django_compressor
djangorestframework
asgiref==3.8.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
channels_redis
channels
daphne
django_browser_reload
random-word
django_rest_framework
Pillow
wonderwords
django-unfold